Hotfix for toolchain selection issue.
Currently, there is an issue where the consuming MODULE.bazel is not
capable of overriding the version of the toolchain specified in our
MODULE.bazel. Create a FIXME comment in extensions.bzl.

Previously, there was also an issue where selecting any toolchain
version besides 9.x.x caused a failure when setting up the external
repository, because the generated BUILD.bazel wasn't provided the
version of the toolchain it was generating.