Use MSV to select the version of the toolchain
This commit fixes an issue with the bzlmod extension that prevents
consuming MODULE.bazel files from being able to select the version of
the toolchain. Now, the bzlmod extension selects the minimum version
from the modules that have requested the toolchain and uses that for the
system. This allows consumers to upgrade the version of
arm_gnu_toolchain without needing to ugprade their compiler version, as
well.