Fix various custom toolchain issues
Previously, there was a change that required the version of the
toolchain to be specified in the cc_arm_*_config rules. This isn't
ideal, because it requires consumers to list the version in two places.
So instead, introduce some new filegroup targets in the cross-hosted
toolchain repository that allows these rules to infer the toolchain
version. Also fix some other existing issues with the examples.