Use rules_cc for many definitions previously provided by bazel_tools
This commit gets us almost all the way to using rules_cc instead of
bazel_tools for forward compatibility after the starlarkification of the
C++ rules is complete. The only import that can't currently be changes
is ACTION_NAMES, because ACTION_NAMES.llvm_cov is not defined in
rules_cc as of today.