Enable support for other ARM GNU Toolchains, add arm-none-linux-gnueabihf toolchain (#30)
* Manage 'target constraints' instead of just compatible cpus

Previously, there was a list of compatible CPUs that were used to form
the target constraints for the toolchain. Now, however, we need to
distinguish toolchains on more than just @platforms//cpu, so manage
these as a collection of constraint values instead.

* Rename module to arm_gnu_toolchain

This commit renames the module to arm_gnu_toolchain from arm_none_eabi.
API-compatibility is _mostly_ retained through the new top-level
repository generation mechanism. Using this module to configure an
arm-none-eabi toolchain results in the generation of an external
repository called @arm_none_eabi. In this way, clients can still rely on
tools and targets that were previously defined by the old module.
Consumers must now import module extensions and repository rules from
the new module, however. See the updated README.md for more information.

* Custom repo rule to templatize compiler BUILD

Introduce a new custom repository rule to allow us to templatize the
"compiler" BUILD.bazel file (toolchain/compilers{nix,win}.BUILD). This
allows us an easy path to changing out the toolchain prefix when
integrating new toolchain types.

* Make cc_arm_none_eabi_config generic

We now encode all toolchain-specific items in the attributes of the
generic "cc_arm_gnu_toolchain_config" rule. This allows us to use this
rule for other ARM GNU toolchains in the future.

* Abstract generic logic from arm_none_eabi_toolchain

All the interesting stuff now lives in a "private" _arm_gnu_toolchain
function that specific toolchains can delegate to. This provide a common
toolchain mechanism that can be used for creating and registering arm
linux toolchains, for example.

* Genericize arm_none_eabi_deps and repo BUILD templates

Continuing with the trend of "genericization," this commit abstracts
everything specific to the arm-none-eabi platform out of the
arm_none_eabi_deps function into a new function that can be leveraged by
additional toolchains. The arm_none_eabi_deps function is kept to
maintain API stability.

* Rename module extension to arm_toolchain and tag classes

Previously, there was a single tag class for the "arm_gnu" module
extension which would be used by consumers to pull in arm vendored
toolchains. Now, however, the tags map to the particular toolchain
through the toolchain's target name. See the updated README for more
information.

* Introduce the arm-none-linux-gnueabihf toolchain

This commit introduces the named toolchain, which functions much in the
same way as the existing arm-none-eabi toolchains. A couple of minor
changes to existing common code are required to get simple
applications to compile.

* Create an example using arm-none-linux-gnueabihf toolchain

* Fix issue that caused tools to be resolved incorrectly

Previously, the BUILD template for the cross-hosted toolchain external
repository (host-specific) had a defect that caused the paths of tools
to be resolved incorrectly.

* Make toolchain rules public

* Hotfix for toolchain selection issue.

Currently, there is an issue where the consuming MODULE.bazel is not
capable of overriding the version of the toolchain specified in our
MODULE.bazel. Create a FIXME comment in extensions.bzl.

Previously, there was also an issue where selecting any toolchain
version besides 9.x.x caused a failure when setting up the external
repository, because the generated BUILD.bazel wasn't provided the
version of the toolchain it was generating.

* Use MSV to select the version of the toolchain

This commit fixes an issue with the bzlmod extension that prevents
consuming MODULE.bazel files from being able to select the version of
the toolchain. Now, the bzlmod extension selects the minimum version
from the modules that have requested the toolchain and uses that for the
system. This allows consumers to upgrade the version of
arm_gnu_toolchain without needing to ugprade their compiler version, as
well.

* Consolidate templates for generated repos

This commit combines the two previous BUILD templates for the generated
repos, nix.BUILD and win.BUILD, into one templatized compiler.BUILD
template, and it also moves the other existing templates into the
toolchain/templates folder.

* Fix various custom toolchain issues

Previously, there was a change that required the version of the
toolchain to be specified in the cc_arm_*_config rules. This isn't
ideal, because it requires consumers to list the version in two places.
So instead, introduce some new filegroup targets in the cross-hosted
toolchain repository that allows these rules to infer the toolchain
version. Also fix some other existing issues with the examples.

* Slightly reorganize examples

Organize the examples by toolchain so that CI workflows can selectively
build examples based on which toolchains are available for the host
platform.

* Fix linux toolchain for mingw-i686

* CI: use --output_user_root to avoid max path length limitation

Previously, CI builds were failing for Windows because the arm-linux
toolchain include paths exceeded the Maximum Path Length. Use
--output_user_root in the workflow to avoid this limitation.

* Use rules_cc for many definitions previously provided by bazel_tools

This commit gets us almost all the way to using rules_cc instead of
bazel_tools for forward compatibility after the starlarkification of the
C++ rules is complete. The only import that can't currently be changes
is ACTION_NAMES, because ACTION_NAMES.llvm_cov is not defined in
rules_cc as of today.

* Document Windows path length issues in README.md

* examples/arm-linux: Link to libc and libstdc++ by default

* non-functional: Bump arm-none-eabi version to 13.2.1 in a few places

This affects default arguments and the README, but should not have any
effect on consumers using older toolchains.

* arm-linux: Default include_std to true