allow toolchains to always link specific libraries (#71)
* allow toolchains to always link specific libraries

Define toolchain attribute `additional_link_libraries` which specifies
libraries that are always linked to an executable. This can be used to
associate specific libraries (e.g. startup code, periph IO) to a
toolchain, removing the need to specify `deps` with `select` when
working with multiple toolchains.

When defining a toolchain, a bootstrap toolchain is created if
`additional_link_libraries` is not empty. This bootstrap toolchain is
used to build all the targets in `additional_link_libraries` and these
built libraries are added as a dependency of the primary toolchain.

* refactor cortex-m examples

* add test with -Wpedantic